Item system rework + enemy class implementation
Rewrote the item system to be stack allocated and only created + deleted when necessary. Also implemented more functionallity for given items.

Enemy class implemented allowing for more overall functionallity per item and character. This is purely for item reasons and the enemy interactions will not be implemeneted, however it would be possible at a later date.